Implement basic delay support
The ini file now has two entries for each corner, one for the action to
be performed and one for the interval after which it will be performed
if the mouse doesn't leave the hot corner/edge. This is a little
inelegant, and I'd like to make the ini file nicer at some point.